+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -6,10 +6,10 @@ CHANGELOG
 
  * added Symfony\Component\HttpKernel\EventListener\RouterProxyListener
  * added Symfony\Component\HttpKernel\UriSigner
- * added Symfony\Component\HttpKernel\HttpContentRenderer and rendering strategies (in Symfony\Component\HttpKernel\RenderingStrategy)
+ * added Symfony\Component\HttpKernel\SubRequestRenderer and rendering strategies (in Symfony\Component\HttpKernel\RenderingStrategy)
  * added Symfony\Component\HttpKernel\EventListener\RouterProxyListener
  * added Symfony\Component\HttpKernel\DependencyInjection\ContainerAwareHttpKernel
- * added ControllerReference to create reference of Controllers (used in the HttpContentRenderer class)
+ * added ControllerReference to create reference of Controllers (used in the SubRequestRenderer class)
  * [BC BREAK] renamed TimeDataCollector::getTotalTime() to
    TimeDataCollector::getDuration()
  * updated the MemoryDataCollector to include the memory used in the